last update commit

+++ api/Controllers/StockController.cs
@@ -4,8 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -25,11 +27,16 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        [Authorize]
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            // var stocks = await _context.Stocks.ToListAsync();
-            var stocks = await _stockRepo.GetAllAsync();
-            var stockDto = stocks.Select(s => s.ToStockDto());
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var stocks = await _stockRepo.GetAllAsync(query);
+            var stockDto = stocks.Select(s => s.ToStockDto()).ToList();
             //select works somewhat similar to filter in javascript, also in this line we are filtering out the attributes/data retrived from database to send in response header
 
             return Ok(stockDto);
@@ -38,7 +45,11 @@ namespace api.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
-            // var stock = await _context.Stocks.FindAsync(id);
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if(stock==null){
@@ -51,9 +62,13 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = stockDto.ToStockFromCreateDTO();
-            // await _context.Stocks.AddAsync(stockModel);
-            // await _context.SaveChangesAsync();
+
             await _stockRepo.CreateAsync(stockModel);
 
             return CreatedAtAction(nameof(GetById), new {id = stockModel.Id}, stockModel.ToStockDto());
@@ -63,22 +78,17 @@ namespace api.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
-            // var stockModel = await _context.Stocks.FirstOrDefaultAsync(x => x.Id == id);
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if(stockModel == null)
             {
                 return NotFound();
             }
-
-            // stockModel.Symbol = updateDto.Symbol;
-            // stockModel.CompanyName = updateDto.CompanyName;
-            // stockModel.Purchase =   updateDto.Purchase;
-            // stockModel.LastDiv = updateDto.LastDiv;
-            // stockModel.Industry = updateDto.Industry;
-            // stockModel.MarketCap =  updateDto.MarketCap;
-            
-            // await _context.SaveChangesAsync();
 
             return Ok(stockModel.ToStockDto());
         }
@@ -88,15 +98,17 @@ namespace api.Controllers
 
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            // var stockModel = await _context.Stocks.FirstOrDefaultAsync(x => x.Id == id);
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null)
             {
                 return NotFound();
             }
-            // _context.Stocks.Remove(stockModel);  // remove/delete operation in dotnet is not a async operation
-            // await _context.SaveChangesAsync();
 
             return NoContent();
         }
